remove go19 support
